requirements.txt
streamlit
youtube-transcript-api
sentence-transformers
faiss-cpu
pandas
numpy
